add feature flag in selectors
