+++ packages/twenty-front/src/modules/prefetch/states/selector/prefetchViewFromViewIdFamilySelector.ts
@@ -1,7 +1,11 @@
+import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { prefetchViewsState } from '@/prefetch/states/prefetchViewsState';
 import { coreViewsState } from '@/views/states/coreViewState';
 import { type View } from '@/views/types/View';
 import { convertCoreViewToView } from '@/views/utils/convertCoreViewToView';
+import { extractFeatureFlagMapFromWorkspace } from '@/workspace/utils/extractFeatureFlagMapFromWorkspace';
 import { selectorFamily } from 'recoil';
+import { FeatureFlagKey } from '~/generated/graphql';
 
 export const prefetchViewFromViewIdFamilySelector = selectorFamily<
   View | undefined,
@@ -11,9 +15,19 @@ export const prefetchViewFromViewIdFamilySelector = selectorFamily<
   get:
     ({ viewId }) =>
     ({ get }) => {
+      const prefetchedViews = get(prefetchViewsState);
       const coreViews = get(coreViewsState);
 
-      const views = coreViews.map(convertCoreViewToView);
+      const currentWorkspace = get(currentWorkspaceState);
+      const featureFlags = extractFeatureFlagMapFromWorkspace(currentWorkspace);
+
+      const isCoreViewSyncingEnabled =
+        featureFlags[FeatureFlagKey.IS_CORE_VIEW_SYNCING_ENABLED];
+
+      const views = isCoreViewSyncingEnabled
+        ? coreViews.map(convertCoreViewToView)
+        : prefetchedViews;
+
       return views?.find((view) => view.id === viewId);
     },
 });
