+++ packages/twenty-front/src/modules/prefetch/states/selector/prefetchViewIdsFromObjectMetadataItemFamilySelector.ts
@@ -1,6 +1,10 @@
+import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { prefetchViewsState } from '@/prefetch/states/prefetchViewsState';
 import { coreViewsState } from '@/views/states/coreViewState';
 import { convertCoreViewToView } from '@/views/utils/convertCoreViewToView';
+import { extractFeatureFlagMapFromWorkspace } from '@/workspace/utils/extractFeatureFlagMapFromWorkspace';
 import { selectorFamily } from 'recoil';
+import { FeatureFlagKey } from '~/generated/graphql';
 
 export const prefetchViewIdsFromObjectMetadataItemFamilySelector =
   selectorFamily<string[], { objectMetadataItemId: string }>({
@@ -8,9 +12,20 @@ export const prefetchViewIdsFromObjectMetadataItemFamilySelector =
     get:
       ({ objectMetadataItemId }) =>
       ({ get }) => {
+        const prefetchedViews = get(prefetchViewsState);
         const coreViews = get(coreViewsState);
 
-        const views = coreViews.map(convertCoreViewToView);
+        const currentWorkspace = get(currentWorkspaceState);
+        const featureFlags =
+          extractFeatureFlagMapFromWorkspace(currentWorkspace);
+
+        const isCoreViewSyncingEnabled =
+          featureFlags[FeatureFlagKey.IS_CORE_VIEW_SYNCING_ENABLED];
+
+        const views = isCoreViewSyncingEnabled
+          ? coreViews.map(convertCoreViewToView)
+          : prefetchedViews;
+
         return views
           .filter((view) => view.objectMetadataId === objectMetadataItemId)
           .map((view) => view.id);
