+++ packages/twenty-front/src/modules/prefetch/states/selector/prefetchViewsFromObjectMetadataItemFamilySelector.ts
@@ -1,7 +1,11 @@
+import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { prefetchViewsState } from '@/prefetch/states/prefetchViewsState';
 import { coreViewsState } from '@/views/states/coreViewState';
 import { type View } from '@/views/types/View';
 import { convertCoreViewToView } from '@/views/utils/convertCoreViewToView';
+import { extractFeatureFlagMapFromWorkspace } from '@/workspace/utils/extractFeatureFlagMapFromWorkspace';
 import { selectorFamily } from 'recoil';
+import { FeatureFlagKey } from '~/generated/graphql';
 
 export const prefetchViewsFromObjectMetadataItemFamilySelector = selectorFamily<
   View[],
@@ -11,9 +15,18 @@ export const prefetchViewsFromObjectMetadataItemFamilySelector = selectorFamily<
   get:
     ({ objectMetadataItemId }) =>
     ({ get }) => {
+      const prefetchedViews = get(prefetchViewsState);
       const coreViews = get(coreViewsState);
 
-      const views = coreViews.map(convertCoreViewToView);
+      const currentWorkspace = get(currentWorkspaceState);
+      const featureFlags = extractFeatureFlagMapFromWorkspace(currentWorkspace);
+
+      const isCoreViewSyncingEnabled =
+        featureFlags[FeatureFlagKey.IS_CORE_VIEW_SYNCING_ENABLED];
+
+      const views = isCoreViewSyncingEnabled
+        ? coreViews.map(convertCoreViewToView)
+        : prefetchedViews;
 
       const filteredViews = views.filter(
         (view) => view.objectMetadataId === objectMetadataItemId,
