+++ packages/twenty-front/src/modules/workspace/hooks/useFeatureFlagsMap.ts
@@ -1,24 +1,10 @@
 import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { extractFeatureFlagMapFromWorkspace } from '@/workspace/utils/extractFeatureFlagMapFromWorkspace';
 import { useRecoilValue } from 'recoil';
-import { FeatureFlagKey } from '~/generated/graphql';
-import { buildRecordFromKeysWithSameValue } from '~/utils/array/buildRecordFromKeysWithSameValue';
+import { type FeatureFlagKey } from '~/generated/graphql';
 
 export const useFeatureFlagsMap = (): Record<FeatureFlagKey, boolean> => {
   const currentWorkspace = useRecoilValue(currentWorkspaceState);
 
-  const currentWorkspaceFeatureFlags = currentWorkspace?.featureFlags;
-
-  const initialFeatureFlags = buildRecordFromKeysWithSameValue(
-    Object.values(FeatureFlagKey),
-    false,
-  );
-
-  if (!currentWorkspaceFeatureFlags) {
-    return initialFeatureFlags;
-  }
-
-  return currentWorkspaceFeatureFlags.reduce((acc, featureFlag) => {
-    acc[featureFlag.key] = featureFlag.value;
-    return acc;
-  }, initialFeatureFlags);
+  return extractFeatureFlagMapFromWorkspace(currentWorkspace);
 };
